Move enums
Make more sense to have them outside of preferences, but in their component/helper
