+++ src/components/SettingsAction.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { ActionKeys, RSPrefs } from "@/preferences";
+import { RSPrefs } from "@/preferences";
 import Locale from "../resources/locales/en.json";
 
 import settingsStyles from "./assets/styles/readerSettings.module.css";
@@ -12,7 +12,7 @@ import CloseIcon from "./assets/icons/close.svg";
 import { Button, Dialog, DialogTrigger, Heading, Popover, Separator } from "react-aria-components";
 import { ActionIcon } from "./Templates/ActionIcon";
 import { OverflowMenuItem } from "./Templates/OverflowMenuItem";
-import { ActionComponentVariant, IActionComponent } from "./Templates/ActionComponent";
+import { ActionComponentVariant, ActionKeys, IActionComponent } from "./Templates/ActionComponent";
 import { ReadingDisplayCol } from "./ReadingDisplayCol";
 import { ReadingDisplayLayout } from "./ReadingDisplayLayout";
