+++ src/helpers/scrollAffordance.ts
@@ -1,5 +1,18 @@
-import { RSPrefs, ScrollAffordancePref } from "@/preferences";
+import { RSPrefs } from "@/preferences";
 import Locale from "../resources/locales/en.json";
+
+export enum ScrollAffordancePref {
+  none = "none",
+  prev = "previous",
+  next = "next",
+  both = "both"
+}
+
+export enum ScrollBackTo {
+  top = "top",
+  bottom = "bottom",
+  untouched = "untouched"
+}
 
 export interface IScrollAffordanceConfig {
   pref: ScrollAffordancePref;
